# internal/httpserve/handlers/register.go
package handlers

import (
	"context"
	"errors"
	"strings"
	"time"

	echo "github.com/theopenlane/echox"

	"github.com/theopenlane/utils/rout"

	"github.com/theopenlane/core/internal/ent/generated"
	"github.com/theopenlane/core/internal/ent/privacy/token"
	entval "github.com/theopenlane/core/internal/ent/validator"
	"github.com/theopenlane/core/internal/integrations/definitions/email"

	"github.com/theopenlane/core/common/enums"
	models "github.com/theopenlane/core/common/openapi"
	"github.com/theopenlane/core/pkg/logx"
)

const (
	maxEmailAttempts = 5
)

// RegisterHandler handles the registration of a new user, creating the user, personal organization
// and sending an email verification to the email address in the request
// the user will not be able to authenticate until the email is verified
func (h *Handler) RegisterHandler(ctx echo.Context) error {
	req, err := BindAndValidate[models.RegisterRequest](ctx)
	if err != nil {
		return h.InvalidInput(ctx, err)
	}

	// create user
	input := generated.CreateUserInput{
		FirstName:         &req.FirstName,
		LastName:          &req.LastName,
		Email:             req.Email,
		Password:          &req.Password,
		LastLoginProvider: &enums.AuthProviderCredentials,
	}

	reqCtx := ctx.Request().Context()

	// set viewer context
	ctxWithToken := token.NewContextWithSignUpToken(reqCtx, req.Email)

	if req.Token != nil {
		ctxWithToken = token.NewContextWithOrgInviteToken(ctxWithToken, *req.Token)

		invitedUser, err := h.getUserByInviteToken(ctxWithToken, *req.Token)
		if err != nil {
			logx.FromContext(reqCtx).Error().Err(err).Str("email", input.Email).Msg("error retrieving invite token")

			return h.BadRequest(ctx, err)
		}

		if !strings.EqualFold(invitedUser.Recipient, input.Email) {
			return h.BadRequest(ctx, ErrUnableToVerifyEmail)
		}

		if invitedUser.Expires.Before(time.Now()) {
			return h.BadRequest(ctx, ErrExpiredToken)
		}
	}

	meowuser, err := h.createUser(ctxWithToken, input)
	if err != nil {
		logx.FromContext(reqCtx).Error().Err(err).Str("email", input.Email).Msg("error creating new user")

		if IsUniqueConstraintError(err) {
			return h.Conflict(ctx, "user already exists", UserExistsErrCode)
		}

		if generated.IsValidationError(err) {
			return h.InvalidInput(ctx, invalidInputError(err))
		}

		if errors.Is(err, entval.ErrEmailNotAllowed) {
			logx.FromContext(reqCtx).Error().Err(err).Str("email", input.Email).Msg("email not allowed")

			return h.InvalidInput(ctx, err)
		}

		return err
	}

	// setup user context
	userCtx := setAuthenticatedContext(ctxWithToken, meowuser)

	if req.Token != nil {
		ctx.SetRequest(ctx.Request().WithContext(userCtx))

		_, _, _, err := h.processInvitation(ctx, *req.Token, meowuser.Email)
		if err != nil {
			logx.FromContext(reqCtx).Error().Err(err).Str("email", input.Email).Msg("error processing invitation")

			return h.BadRequest(ctx, ErrUnableToVerifyEmail)
		}

		if err := h.setEmailConfirmed(userCtx, meowuser); err != nil {
			logx.FromContext(reqCtx).Error().Err(err).Str("email", input.Email).Msg("unable to set email as verified")

			return h.BadRequest(ctx, ErrUnableToVerifyEmail)
		}
	}

	out := &models.RegisterReply{
		Reply:   rout.Reply{Success: true},
		ID:      meowuser.ID,
		Email:   meowuser.Email,
		Message: "Welcome to Openlane!",
	}

	if req.Token == nil {
		// create email verification token
		user := &User{
			FirstName: req.FirstName,
			LastName:  req.LastName,
			Email:     req.Email,
			ID:        meowuser.ID,
		}

		meowtoken, err := h.storeAndSendEmailVerificationToken(userCtx, user)
		if err != nil {
			logx.FromContext(reqCtx).Error().Err(err).Str("email", input.Email).Msg("error storing email verification token")

			return h.InternalServerError(ctx, ErrProcessingRequest)
		}

		// only return the token in development
		if h.IsDev {
			out.Token = meowtoken.Token
		}
	}

	return h.Created(ctx, out)
}

func (h *Handler) storeAndSendEmailVerificationToken(ctx context.Context, user *User) (*generated.EmailVerificationToken, error) {
	// expire all existing tokens
	if err := h.expireAllVerificationTokensUserByEmail(ctx, user.Email); err != nil {
		logx.FromContext(ctx).Error().Err(err).Str("email", user.Email).Msg("error expiring existing tokens")

		return nil, err
	}

	// check if the user has attempted to verify their email too many times
	attempts, err := h.countVerificationTokensUserByEmail(ctx, user.Email)
	if err != nil {
		logx.FromContext(ctx).Error().Err(err).Str("email", user.Email).Msg("error getting existing tokens")

		return nil, err
	}

	if attempts >= maxEmailAttempts {
		return nil, ErrMaxAttempts
	}

	// create a new token and store it in the database
	if err := user.CreateVerificationToken(); err != nil {
		logx.FromContext(ctx).Error().Err(err).Str("email", user.Email).Msg("error creating verification token")

		return nil, err
	}

	meowtoken, err := h.createEmailVerificationToken(ctx, user)
	if err != nil {
		return nil, err
	}

	if err := h.sendEmail(ctx, email.VerifyEmailOp.Name(), email.VerifyEmailRequest{
		RecipientInfo: email.RecipientInfo{Email: user.Email, FirstName: user.FirstName, LastName: user.LastName},
		Token:         meowtoken.Token,
	}); err != nil {
		return nil, err
	}

	return meowtoken, nil
}