# internal/httpserve/specs/openlane.openapi.json
{
  "components": {
    "parameters": {
      "attributes": {
        "description": "A comma-separated list of attribute names to return in the response,\noverriding the default set of attributes. Attribute names must be in\nstandard attribute notation form.\n",
        "example": "userName,emails,name",
        "in": "query",
        "name": "attributes",
        "schema": {
          "type": "string"
        }
      },
      "count": {
        "description": "The number of resources returned in a list response page",
        "in": "query",
        "name": "count",
        "schema": {
          "default": 100,
          "minimum": 1,
          "type": "integer"
        }
      },
      "excludedAttributes": {
        "description": "A comma-separated list of attribute names to exclude from the default\nset of attributes to return. Attribute names must be in standard\nattribute notation form.\n",
        "example": "groups",
        "in": "query",
        "name": "excludedAttributes",
        "schema": {
          "type": "string"
        }
      },
      "filter": {
        "description": "Filter expression for querying resources. Follows SCIM filter syntax as defined in\n[RFC 7644 Section 3.4.2.2](https://datatracker.ietf.org/doc/html/rfc7644#section-3.4.2.2).\n\nExample: `userName eq \"mitb@meowmeow.com\"`\n",
        "in": "query",
        "name": "filter",
        "schema": {
          "type": "string"
        }
      },
...
      "startIndex": {
        "description": "The 1-based index of the first result in the current set of list results",
        "in": "query",
        "name": "startIndex",
        "schema": {
          "default": 1,
          "minimum": 1,
          "type": "integer"
        }
      },
...
    },
    "requestBodies": {
...
      "UserBody": {
        "content": {
          "application/scim+json": {
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        },
        "description": "User resource attributes",
        "required": true
      }
    },
    "responses": {
      "BadRequest": {
        "content": {
          "application/scim+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Request is unparsable, syntactically incorrect, or violates schema"
      },
      "Conflict": {
        "content": {
          "application/scim+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "The specified version number does not match the resource's latest version number,\nor a service provider refused to create a new, duplicate resource\n"
      },
      "Forbidden": {
        "content": {
          "application/scim+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Operation is not permitted based on the supplied authorization"
      },
...
      "InternalServerError": {
        "content": {
          "application/scim+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Internal Server Error"
      },
...
      "Unauthorized": {
        "content": {
          "application/scim+json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Authorization failure. The authorization header is invalid or missing"
      },
      "UserCreated": {
        "content": {
          "application/scim+json": {
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        },
        "description": "User created successfully",
        "headers": {
          "Location": {
            "description": "The URL of the newly created user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "UserList": {
        "content": {
          "application/scim+json": {
            "schema": {
              "properties": {
                "Resources": {
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "type": "array"
                },
                "itemsPerPage": {
                  "$ref": "#/components/schemas/itemsPerPage"
                },
                "schemas": {
                  "$ref": "#/components/schemas/listResponseSchemas"
                },
                "startIndex": {
                  "$ref": "#/components/schemas/startIndex"
                },
                "totalResults": {
                  "$ref": "#/components/schemas/totalResults"
                }
              },
              "type": "object"
            }
          }
        },
        "description": "List of users"
      },
...
    },
    "schemas": {
...
      "ErrorReply": {
        "properties": {
          "error": {
            "description": "The error message if the request was unsuccessful",
            "type": "string"
          },
          "error_code": {
            "description": "The error code if the request was unsuccessful",
            "type": "string"
          },
          "success": {
            "description": "Whether or not the request was successful or not",
            "type": "boolean"
          },
          "unverified": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Error response as defined in RFC 7644 Section 3.12",
        "properties": {
          "detail": {
            "description": "A detailed human-readable message",
            "type": "string"
          },
          "schemas": {
            "items": {
              "enum": [
                "urn:ietf:params:scim:api:messages:2.0:Error"
              ],
              "type": "string"
            },
            "type": "array"
          },
          "scimType": {
            "description": "A SCIM detail error type",
            "enum": [
              "invalidFilter",
              "tooMany",
              "uniqueness",
              "mutability",
              "invalidSyntax",
              "invalidPath",
              "noTarget",
              "invalidValue",
              "invalidVers",
              "sensitive"
            ],
            "type": "string"
          },
          "status": {
            "description": "The HTTP status code expressed as a JSON string",
            "enum": [
              "400",
              "401",
              "403",
              "404",
              "409",
              "500",
              "501"
            ],
            "type": "string"
          }
        },
        "required": [
          "schemas",
          "status"
        ],
        "type": "object"
      },
...
      "GetQuestionnaireResponse": {
        "properties": {
          "jsonconfig": {
            "additionalProperties": {},
            "type": "object"
          },
          "saved_data": {
            "additionalProperties": {},
            "type": "object"
          },
          "uischema": {
            "additionalProperties": {},
            "type": "object"
          }
        },
        "type": "object"
      },
...
      "ResendQuestionnaireRequest": {
        "properties": {
          "assessment_id": {
            "description": "The ID of the assessment",
            "type": "string"
          },
          "email": {
            "description": "The email address associated with the assessment response",
            "type": "string"
          }
        },
        "type": "object"
      },
      "ResendReply": {
        "properties": {
          "error": {
            "description": "The error message if the request was unsuccessful",
            "type": "string"
          },
          "error_code": {
            "description": "The error code if the request was unsuccessful",
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "description": "Whether or not the request was successful or not",
            "type": "boolean"
          },
          "unverified": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "SubmitQuestionnaireRequest": {
        "properties": {
          "assessment_id": {
            "type": "string"
          },
          "data": {
            "additionalProperties": {},
            "type": "object"
          },
          "is_draft": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "SubmitQuestionnaireResponse": {
        "properties": {
          "completed_at": {
            "type": "string"
          },
          "document_data_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "itemsPerPage": {
        "description": "The number of resources returned in the current page",
        "minimum": 0,
        "type": "integer"
      },
      "listResponseSchemas": {
        "items": {
          "enum": [
            "urn:ietf:params:scim:api:messages:2.0:ListResponse"
          ],
          "type": "string"
        },
        "type": "array"
      },
      "startIndex": {
        "description": "The 1-based index of the first result in the current set of results",
        "minimum": 1,
        "type": "integer"
      },
      "totalResults": {
        "description": "The total number of results matching the query",
        "minimum": 0,
        "type": "integer"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Bearer token authentication using API tokens.\nThe token identifies the organization and grants permissions to manage\nusers and groups within that organization's scope.\n",
        "scheme": "bearer",
        "type": "http"
      },
      "apiKey": {
        "description": "API Key authentication, the key must be a valid API key",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "bearer": {
        "bearerFormat": "JWT",
        "description": "Bearer authentication, the token must be a valid API token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
...
  "info": {
    "contact": {
      "email": "support@theopenlane.io",
      "name": "Openlane",
      "url": "https://www.theopenlane.io"
    },
    "description": "Openlane's API services are designed to provide a simple and easy to use interface for interacting with the Openlane platform. This API is designed to be used by both internal and external clients to interact with the Openlane platform.",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title": "Openlane OpenAPI 3.1.1 Specifications",
    "version": "v0.0.0-dev"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/questionnaire": {
      "get": {
        "description": "Get questionnaire template configuration for authenticated anonymous users",
        "operationId": "GetQuestionnaire",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "GetQuestionnaireResponse": {
                    "value": {
                      "jsonconfig": {
                        "description": "A sample questionnaire template",
                        "title": "Sample Questionnaire"
                      }
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/GetQuestionnaireResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "bad request",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "unauthorized",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "not found",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "internal server error",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "apiKey": []
          }
        ],
        "summary": "GetQuestionnaire",
        "tags": [
          "Questionnaires"
        ]
      },
      "post": {
        "description": "Submit questionnaire response data for authenticated anonymous users",
        "operationId": "SubmitQuestionnaire",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "SubmitQuestionnaireRequest": {
                  "value": {
                    "data": {
                      "q1": "Answer to question 1",
                      "q2": "Answer to question 2",
                      "q3": {
                        "nested": "data"
                      }
                    }
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/SubmitQuestionnaireRequest"
              }
            }
          },
          "description": "Request body"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "SubmitQuestionnaireResponse": {
                    "value": {
                      "completed_at": "2024-01-01T12:00:00Z",
                      "document_data_id": "01JCQR8Z9X1A2B3C4D5E6F7G8H",
                      "status": "COMPLETED"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/SubmitQuestionnaireResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "bad request",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "unauthorized",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "not found",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "internal server error",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearer": []
          },
          {
            "apiKey": []
          }
        ],
        "summary": "SubmitQuestionnaire",
        "tags": [
          "Questionnaires"
        ]
      }
    },
    "/questionnaire/resend": {
      "post": {
        "description": "Resend questionnaire authentication email with new JWT token",
        "operationId": "ResendQuestionnaireEmail",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "ResendQuestionnaireRequest": {
                  "value": {
                    "assessment_id": "01JCQR8Z9X1A2B3C4D5E6F7G8H",
                    "email": "vendor@example.com"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/ResendQuestionnaireRequest"
              }
            }
          },
          "description": "Request body"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "ResendEmailSuccessResponse": {
                    "value": {
                      "message": "Email has been resent",
                      "success": true
                    }
                  },
                  "ResendQuestionnaireResponse": {
                    "value": {
                      "message": "If the email address is associated with an active assessment, a new authentication email has been sent",
                      "success": true
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ResendReply"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "bad request",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Bad Request"
          },
          "429": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "too many requests",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "error": {
                    "value": {
                      "error": "internal server error",
                      "success": false
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorReply"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [],
        "summary": "ResendQuestionnaireEmail",
        "tags": [
          "Questionnaires"
        ]
      }
    },
...
    "/scim/v2/Users": {
      "get": {
        "description": "Returns a list of users with optional filtering and pagination",
        "operationId": "listUsers",
        "parameters": [
          {
            "$ref": "#/components/parameters/filter"
          },
          {
            "$ref": "#/components/parameters/startIndex"
          },
          {
            "$ref": "#/components/parameters/count"
          },
          {
            "$ref": "#/components/parameters/attributes"
          },
          {
            "$ref": "#/components/parameters/excludedAttributes"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/UserList"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List users",
        "tags": [
          "SCIM Users"
        ]
      },
      "post": {
        "description": "Creates a new user with the provided attributes",
        "operationId": "createUser",
        "requestBody": {
          "$ref": "#/components/requestBodies/UserBody"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/UserCreated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create a new user",
        "tags": [
          "SCIM Users"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Openlane API Server",
      "url": "https://api.theopenlane.io"
    }
  ],
...
}